+++ internal/numeric.h
@@ -35,9 +35,18 @@ enum ruby_num_rounding_mode {
     RUBY_NUM_ROUND_DEFAULT = ROUND_DEFAULT,
 };
 
+#if SIZEOF_DOUBLE <= SIZEOF_VALUE
+typedef double rb_float_value_type;
+#else
+typedef struct {
+    VALUE values[(SIZEOF_DOUBLE + SIZEOF_VALUE - 1) / SIZEOF_VALUE];
+    /* roomof() needs internal.h, and the order of some macros may matter */
+} rb_float_value_type;
+#endif
+
 struct RFloat {
     struct RBasic basic;
-    double float_value;
+    rb_float_value_type float_value;
 };
 
 #define RFLOAT(obj)  ((struct RFloat *)(obj))
@@ -209,7 +218,15 @@ rb_float_flonum_value(VALUE v)
 static inline double
 rb_float_noflonum_value(VALUE v)
 {
+#if SIZEOF_DOUBLE <= SIZEOF_VALUE
     return RFLOAT(v)->float_value;
+#else
+    union {
+        rb_float_value_type v;
+        double d;
+    } u = {RFLOAT(v)->float_value};
+    return u.d;
+#endif
 }
 
 static inline double